requirements.txt
fastapi
tortoise-orm
passlib
python-jose
uvicorn
pydantic[email]